refactor: 상수 폴더 위치 수정
